Don't try to move files that don't exist

+++ make_links.sh
@@ -5,7 +5,9 @@ THIS_DIR="$(dirname $0)"
 THIS_DIR_RELATIVE="$(realpath --relative-to="$HOME" "$THIS_DIR")"
 
 for file in .bashrc .bash_aliases .vimrc .tmux.conf .tmux_powerline.conf .gitconfig; do
-	mv "$HOME/$file" "$HOME/$file.old"
+	if [ -e "$HOME/$file" ]; then
+		mv "$HOME/$file" "$HOME/$file.old"
+	fi
 	ln -s "$THIS_DIR_RELATIVE/$file" "$HOME/$file"
 done
